Plotting buy/ sell signals with a simple moving average
If prices are higher in the short-term than in the long-term then the current price is relatively high,  indicating short-term price momentum in the upwards direction. Upwards price momentum is classified as a buy signal and downwards momentum is classified as a sell signal. The trading system would buy/ sell accordingly.